Split build and create integration tests
