+++ test/data/definition_files/no_galaxy.yml
@@ -1,0 +1,4 @@
+---
+version: 1
+dependencies:
+  galaxy: doesnotexist.yml
