+++ test/data/definition_files/no_python.yml
@@ -1,0 +1,4 @@
+---
+version: 1
+dependencies:
+  python: doesnotexist.txt
